import-util/requirements.txt
boto==2.49.0
boto3==1.9.190
botocore==1.12.190
Click==7.0
configparser==3.7.4
docutils==0.14
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
lxml==4.4.0
MarkupSafe==1.1.1
patool==1.12
python-dateutil==2.8.0
python-docx==0.8.10
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.5
